rss indexer fixes + moving indexer settings to main db
