+++ src/Torrential.Extensions.Indexing/Clients/IIndexerClient.cs
@@ -1,4 +1,5 @@
 using Torrential.Extensions.Indexing.Models;
+using Torrential.Models;
 
 namespace Torrential.Extensions.Indexing.Clients;
