+++ src/Torrential.Extensions.Indexing/IndexingExtensions.cs
@@ -1,6 +1,5 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Torrential.Extensions.Indexing.Clients;
@@ -14,17 +13,6 @@ public static class IndexingExtensions
 {
     public static IServiceCollection AddTorrentialIndexing(this IServiceCollection services, IConfiguration configuration)
     {
-        var dbPath = Path.Combine(
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
-            "Torrential", "indexing.db");
-
-        var dbDirectory = Path.GetDirectoryName(dbPath);
-        if (!string.IsNullOrEmpty(dbDirectory) && !Directory.Exists(dbDirectory))
-            Directory.CreateDirectory(dbDirectory);
-
-        services.AddDbContext<IndexingDbContext>(options =>
-            options.UseSqlite($"Data Source={dbPath}"));
-
         services.AddHttpClient("Indexer");
 
         services.AddSingleton<IIndexerRepository, IndexerRepository>();
@@ -43,16 +31,5 @@ public static class IndexingExtensions
     {
         IndexingEndpoints.Map(app);
         return app;
-    }
-
-    /// <summary>
-    /// Ensures the indexing database is created and migrated.
-    /// Call this during application startup.
-    /// </summary>
-    public static async Task InitializeIndexingAsync(this IServiceProvider services)
-    {
-        using var scope = services.CreateScope();
-        var db = scope.ServiceProvider.GetRequiredService<IndexingDbContext>();
-        await db.Database.EnsureCreatedAsync();
     }
 }
