+++ src/Torrential.Web/Api/Requests/Torrents/TorrentAddFromUrlRequest.cs
@@ -1,0 +1,7 @@
+namespace Torrential.Web.Api.Requests.Torrents
+{
+    public class TorrentAddFromUrlRequest
+    {
+        public required string Url { get; init; }
+    }
+}
